# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/carts": {
      "get": {
        "operationId": "CartsController_getCart",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Panier récupéré avec succès"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtenir le panier de l'utilisateur",
        "tags": [
          "Carts - Panier"
        ]
      },
      "delete": {
        "operationId": "CartsController_clearCart",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Panier vidé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Vider le panier",
        "tags": [
          "Carts - Panier"
        ]
      }
    },
    "/carts/count": {
      "get": {
        "operationId": "CartsController_getCartCount",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Nombre d'items récupéré"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtenir le nombre d'items dans le panier",
        "tags": [
          "Carts - Panier"
        ]
      }
    },
    "/carts/items": {
      "post": {
        "operationId": "CartsController_addToCart",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddToCartDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Produit ajouté au panier"
          },
          "400": {
            "description": "Stock insuffisant"
          },
          "404": {
            "description": "Produit non trouvé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ajouter un produit au panier",
        "tags": [
          "Carts - Panier"
        ]
      }
    },
    "/carts/items/{cartItemId}": {
      "put": {
        "operationId": "CartsController_updateCartItem",
        "parameters": [
          {
            "name": "cartItemId",
            "required": true,
            "in": "path",
            "description": "ID de l'item du panier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCartDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item du panier mis à jour"
          },
          "404": {
            "description": "Item du panier non trouvé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mettre à jour un item du panier",
        "tags": [
          "Carts - Panier"
        ]
      },
      "delete": {
        "operationId": "CartsController_removeCartItem",
        "parameters": [
          {
            "name": "cartItemId",
            "required": true,
            "in": "path",
            "description": "ID de l'item du panier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Item du panier supprimé"
          },
          "404": {
            "description": "Item du panier non trouvé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprimer un item du panier",
        "tags": [
          "Carts - Panier"
        ]
      }
    },
    "/carts/validate": {
      "get": {
        "operationId": "CartsController_validateCart",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Validation du panier"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Valider le panier avant checkout",
        "tags": [
          "Carts - Panier"
        ]
      }
    },
    "/carts/checkout": {
      "post": {
        "operationId": "CartsController_checkout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Commande créée avec succès"
          },
          "400": {
            "description": "Panier invalide ou vide"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Passer commande (checkout)",
        "tags": [
          "Carts - Panier"
        ]
      }
    },
...
    "/cache/stats": {
      "get": {
        "description": "Nombre de clés, mémoire utilisée, état de la connexion",
        "operationId": "CacheController_getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Statistiques récupérées"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Statistiques du cache Redis",
        "tags": [
          "cache"
        ]
      }
    },
    "/cache/exists/{key}": {
      "get": {
        "operationId": "CacheController_exists",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "description": "Clé à vérifier",
            "schema": {
              "example": "product:uuid-123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Résultat de la vérification",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "key": {
                      "type": "string"
                    },
                    "exists": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Vérifier si une clé existe",
        "tags": [
          "cache"
        ]
      }
    },
    "/cache/{key}": {
      "delete": {
        "operationId": "CacheController_delete",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "description": "Clé à supprimer",
            "schema": {
              "example": "product:uuid-123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Clé supprimée"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Supprimer une clé du cache",
        "tags": [
          "cache"
        ]
      }
    },
    "/cache/pattern/{path}": {
      "delete": {
        "description": "Supprime toutes les clés correspondant au pattern",
        "operationId": "CacheController_deleteByPattern",
        "parameters": [
          {
            "name": "pattern",
            "required": true,
            "in": "query",
            "description": "Pattern (ex: product:*, user:*)",
            "schema": {
              "example": "product:*",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cache vidé"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Vider le cache par pattern",
        "tags": [
          "cache"
        ]
      }
    },
    "/cache/flush": {
      "delete": {
        "description": "⚠️ Action destructive - Supprime toutes les clés",
        "operationId": "CacheController_flush",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Cache vidé complètement"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Vider tout le cache",
        "tags": [
          "cache"
        ]
      }
    },
    "/cache/keys": {
      "get": {
        "description": "Retourne les clés correspondant au pattern (max 100)",
        "operationId": "CacheController_getKeys",
        "parameters": [
          {
            "name": "pattern",
            "required": false,
            "in": "query",
            "description": "Pattern de recherche",
            "schema": {
              "example": "*",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des clés"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Lister les clés du cache",
        "tags": [
          "cache"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Documentation",
    "description": "API documentation for the application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "JWT-refresh": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Refresh-JWT",
        "description": "Enter refresh JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "AddToCartDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "description": "ID du produit à ajouter au panier",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "quantity": {
            "type": "number",
            "description": "Quantité du produit",
            "minimum": 1,
            "maximum": 99,
            "default": 1
          }
        },
        "required": [
          "productId",
          "quantity"
        ]
      },
      "UpdateCartDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "description": "ID du produit à ajouter au panier",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "quantity": {
            "type": "number",
            "description": "Nouvelle quantité du produit",
            "minimum": 1,
            "maximum": 99,
            "default": 1
          }
        }
      },
      "CheckoutDto": {
        "type": "object",
        "properties": {
          "cartId": {
            "type": "string",
            "description": "ID du panier (si non fourni, utilise le panier actif)",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "discountCode": {
            "type": "string",
            "description": "Code promo (si applicable)",
            "example": "PROMO2024"
          },
          "shippingAddress": {
            "type": "string",
            "description": "Adresse de livraison",
            "example": "123 Rue de Tunis, Tunis 1000"
          },
          "paymentMethod": {
            "type": "string",
            "description": "Méthode de paiement",
            "example": "flouci",
            "enum": [
              "flouci",
              "cash_on_delivery"
            ]
          },
          "notes": {
            "type": "string",
            "description": "Notes supplémentaires",
            "example": "Laisser le colis à la porte"
          }
        },
        "required": [
          "shippingAddress",
          "paymentMethod"
        ]
      },
...
    }
  }
}